perf: C++側のパフォーマンス向上
- Gravityクラスの見直し
- 不必要に小さい整数型の使用を解消

+++ native/include/puyotan/game/tsumo.hpp
@@ -15,19 +15,19 @@ class Tsumo {
 public:
     explicit Tsumo(int32_t seed = 0);
 
-    inline PuyoPiece get(int index) const {
+    inline PuyoPiece get(int32_t index) const {
         if (index >= generated_count_) {
             generateMore();
         }
-        // Ring buffer access using bitmask (size 512 -> mask 0x1FF)
-        return pool_[index & (config::Rule::kTsumoPoolSize - 1)];
+        // Ring buffer access using bitmask
+        return pool_[static_cast<std::size_t>(index) & (config::Rule::kTsumoPoolSize - 1)];
     }
     void setSeed(int32_t seed);
     int32_t getSeed() const { return seed_; }
 
 private:
     mutable int32_t seed_;
-    mutable int generated_count_ = 0;
+    mutable int32_t generated_count_ = 0;
     mutable std::array<PuyoPiece, config::Rule::kTsumoPoolSize> pool_;
 
     static_assert((config::Rule::kTsumoPoolSize & (config::Rule::kTsumoPoolSize - 1)) == 0,
